requirements.txt
asgiref==3.9.1
Django==4.2.16
django-phonenumber-field==8.1.0
phonenumbers==9.0.11
sqlparse==0.5.3
tzdata==2025.2
